alice/paskills/nlu/requirements.txt
tornado==4.3
click==6.7
jinja2==2.10
typing==3.6.4
attrs==17.4.0
enum34==1.1.6
pynorm==0.1.11
yandex-inflector==0.1.1
